+++ samples/src/main/java/Demo5Audio.java
@@ -13,11 +13,11 @@ import dev.gamekit.ui.enums.CrossAxisAlignment;
 import dev.gamekit.ui.enums.MainAxisAlignment;
 import dev.gamekit.ui.widgets.*;
 import dev.gamekit.ui.widgets.Image;
+import dev.gamekit.utils.EngineImage;
 import dev.gamekit.utils.Position;
 import dev.gamekit.utils.Vector;
 
 import java.awt.*;
-import java.awt.image.BufferedImage;
 import java.util.Objects;
 
 /**
@@ -33,7 +33,7 @@ import java.util.Objects;
  * </ul>
  */
 public class Demo5Audio extends Scene {
-  private static final BufferedImage SPEAKER_IMG = IO.getResourceImage("speaker.png");
+  private static final EngineImage SPEAKER_IMG = IO.getImage("speaker.png");
   private static final String MUSIC_KEY = "music";
   private final int halfWindowWidth;
   private final Vector listenerPos;